Add bug fixes from main and upgrade to version 1.0.1 (#60)
* Adding the initial changes for purger fix

* Removing deletionInterval from MetricConfig class

* Moving the removeMetrics() back to PerformanceAnalyzerMetrics class

* Adding Event log file deletion + deletion time metric

* Updating the EventLogFilesDeleted metrics

* Moving METRICS_WRITE_ERROR metric to Named Counter, updating deleteFiles() signature

* nit: Removing an unused import

* Addressing the spotbug failure

* Fixing the linker error, updating from docs-beta to official documentation

* Excluding https://www.envoyproxy.io/ from link checker. Failing with too many redirects

* Create writer file only if metrics are present (#23)

* Create writer file only if metrics are present


* Add metrics to writers metric

* Rename writer_file_creation metric name


* Update the statistic for the metric

* Spotless changes


* Add privileges for removing files (#45)


* Remove spotless dependency on main (#54)

This PR removes the need for the main branch to be checked out
in order to build.


* Upgrade to version 1.0.1

+++ src/main/java/org/opensearch/performanceanalyzer/reader_writer_shared/EventLogFileHandler.java
@@ -39,10 +39,14 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardOpenOption;
+import java.util.Arrays;
 import java.util.List;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.opensearch.performanceanalyzer.PerformanceAnalyzerApp;
 import org.opensearch.performanceanalyzer.core.Util;
+import org.opensearch.performanceanalyzer.metrics.PerformanceAnalyzerMetrics;
+import org.opensearch.performanceanalyzer.rca.framework.metrics.WriterMetrics;
 import org.opensearch.performanceanalyzer.reader.EventDispatcher;
 
 public class EventLogFileHandler {
@@ -78,8 +82,10 @@ public class EventLogFileHandler {
      * data.
      *
      * <p>If any of the above steps fail, then the tmp file is not deleted from the filesystem. This
-     * is fine as the MetricsPurgeActivity, will eventually clean it. The copies are atomic and
-     * therefore the reader never reads incompletely written file.
+     * is fine as the {@link
+     * org.opensearch.performanceanalyzer.reader_writer_shared.EventLogFileHandler#deleteFiles},
+     * will eventually clean it. The copies are atomic and therefore the reader never reads
+     * incompletely written file.
      *
      * @param dataEntries The metrics to be written to file.
      * @param epoch The epoch all the metrics belong to.
@@ -113,11 +119,19 @@ public class EventLogFileHandler {
     public void renameFromTmpWithPrivilege(long epoch) {
         Path path = Paths.get(metricsLocation, String.valueOf(epoch));
         Path tmpPath = Paths.get(path.toString() + TMP_FILE_EXT);
-        // This is done only when no exception is thrown.
-        try {
-            Files.move(tmpPath, path, REPLACE_EXISTING, ATOMIC_MOVE);
-        } catch (IOException e) {
-            LOG.error("Error moving file {} to {}.", tmpPath.toString(), path.toString(), e);
+        File tempFile = new File(tmpPath.toString());
+        // Only if the tmp file is present, we want to create the writer file.
+        // If not, we will publish a metric.
+        if (tempFile.exists()) {
+            // This is done only when no exception is thrown.
+            try {
+                Files.move(tmpPath, path, REPLACE_EXISTING, ATOMIC_MOVE);
+            } catch (IOException e) {
+                LOG.error("Error moving file {} to {}.", tmpPath.toString(), path.toString(), e);
+            }
+        } else {
+            PerformanceAnalyzerApp.WRITER_METRICS_AGGREGATOR.updateStat(
+                    WriterMetrics.WRITER_FILE_CREATION_SKIPPED, "", 1);
         }
     }
 
@@ -131,7 +145,7 @@ public class EventLogFileHandler {
         File tempFile = new File(pathToFile.toString());
         if (!tempFile.exists()) {
             long mCurrT = System.currentTimeMillis();
-            LOG.info("Didnt find {} at {}", filename, mCurrT);
+            LOG.debug("Didnt find {} at {}", filename, mCurrT);
             return;
         }
         readInternal(pathToFile, BUFFER_SIZE, processor);
@@ -165,5 +179,45 @@ public class EventLogFileHandler {
         } catch (IOException ex) {
             LOG.error("Error reading file", ex);
         }
+    }
+
+    public void deleteAllFiles() {
+        Util.invokePrivileged(this::deleteAllFilesWithPrivilege);
+    }
+
+    public void deleteAllFilesWithPrivilege() {
+        LOG.debug("Cleaning up any leftover files in [{}]", metricsLocation);
+        File root = new File(metricsLocation);
+        String[] filesToDelete = root.list();
+        if (filesToDelete == null) {
+            return;
+        }
+        deleteFiles(Arrays.asList(filesToDelete));
+    }
+
+    public void deleteFiles(List<String> filesToDelete) {
+        LOG.debug("Starting to delete old writer files");
+        long startTime = System.currentTimeMillis();
+
+        if (filesToDelete == null) {
+            return;
+        }
+        int filesDeletedCount = 0;
+        File root = new File(metricsLocation);
+        for (String fileToDelete : filesToDelete) {
+            File file = new File(root, fileToDelete);
+            removeFilesWithPrivilege(file);
+            filesDeletedCount += 1;
+        }
+        long duration = System.currentTimeMillis() - startTime;
+        PerformanceAnalyzerApp.WRITER_METRICS_AGGREGATOR.updateStat(
+                WriterMetrics.EVENT_LOG_FILES_DELETION_TIME, "", duration);
+        PerformanceAnalyzerApp.WRITER_METRICS_AGGREGATOR.updateStat(
+                WriterMetrics.EVENT_LOG_FILES_DELETED, "", filesDeletedCount);
+        LOG.debug("'{}' Old writer files cleaned up.", filesDeletedCount);
+    }
+
+    public void removeFilesWithPrivilege(File file) {
+        Util.invokePrivileged(() -> PerformanceAnalyzerMetrics.removeMetrics(file));
     }
 }
